requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-autoslug==1.9.8
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
